+++ render/Main.cpp
@@ -1,0 +1,9 @@
+
+#include"stdafx.h"
+#include<iostream>
+#include"rendering.h"
+void main() {
+	device dev;
+	void *fb;
+//	device_init(&dev,640,480,fb);
+}
