+++ render/stdafx.cpp
@@ -1,0 +1,2 @@
+#include"stdafx.h"
+int CMID(int x, int min, int max) { return(x < min) ? min : ((x > max) ? max : x); }
